requirements.txt
django-unrest-0.1.9
matplotlib==3.5.2
numpy==1.23.1
opencv-python==4.6.0.66
typer==0.6.1
urcv==0.0.6
